Need to make probability calculator
So turns out the data is stored in a rather painful to use manner so that'll also need to be reworked. However right now all I need to do is make it so it can take a decision and from its dataset work out the probability of if it should do that or not.